+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+* Fixed an issue where **validate** with the `--input` flag failed on ignored error.
 * Enable passing a comma-separated list of paths for the `--input` option of the **lint** command.
 * Added new validation of unimplemented test-module command in the code to the `XSOAR-linter` in the **lint** command.
 * Fixed the **generate-docs** to handle integration authentication parameter.