+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Fixed an issue where **download** did not return exit code 1 on failure.
 * Updated the validation that incident fields' names do not contain the word incident will aplly to core packs only.
 * Added a playbook validation to verify all conditional tasks have an 'else' path in **validate** command.
+* Renamed the GitHub authentication token environment variable `GITHUB_TOKEN` to `DEMITO_SDK_GITHUB_TOKEN`.
 * Added to the **update-release-notes** command automatic addition to git when new release notes file is created.
 * Added validation to ensure that integrations, scripts, and playbooks do not contain the entity type in their names.
 * Added the **convert** command to convert entities between XSOAR versions.