+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Added validation to ensure that integrations, scripts, and playbooks do not contain the entity type in their names.
 * Added the **convert** command to convert entities between XSOAR versions.
 * Added the *--deprecate* flag in **format** command to deprecate integrations, scripts, and playbooks.
+* Fixed an issue where ignoring errors did not work when running the **validate** command on specific files (-i).
 
 # 1.3.9
 * Added a validation verifying that the pack's README.md file is not equal to pack description.