+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+
+
+# 1.4.0
 * Enable passing a comma-separated list of paths for the `--input` option of the **lint** command.
 * Added new validation of unimplemented test-module command in the code to the `XSOAR-linter` in the **lint** command.
 * Fixed the **generate-docs** to handle integration authentication parameter.
@@ -12,7 +15,7 @@
 * Updated the **upload** command to support reports.
 * Fixed an issue where **generate-docs** command was displaying 'None' when credentials parameter display field configured was not configured.
 * Fixed an issue where **download** did not return exit code 1 on failure.
-* Updated the validation that incident fields' names do not contain the word incident will aplly to core packs only.
+* Updated the validation that incident fields names do not contain the word incident will aplly to core packs only.
 * Added a playbook validation to verify all conditional tasks have an 'else' path in **validate** command.
 * Renamed the GitHub authentication token environment variable `GITHUB_TOKEN` to `DEMITO_SDK_GITHUB_TOKEN`.
 * Added to the **update-release-notes** command automatic addition to git when new release notes file is created.